requirements.txt
flask>=2.3.0
Pillow>=10.0.0
Flask-SQLAlchemy>=3.0.0
Flask-Login>=0.6.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
